complete

+++ backend/routes/userCRUD/delete.js
@@ -4,10 +4,15 @@ const User=require('../../models/users')
 
 router.post("/", async (req, res) => {
     try{
-        const {email}=req.body
+        const {email,admin}=req.body
         const user=await User()
-        const result=await user.deleteOne({email:email})
-        res.status(200).json(result)
+        if (admin===true){
+            const result=await user.deleteOne({email:email})
+            res.status(200).json(result)
+        }
+        else{
+            res.status(401).json({ error: "You are not authorized to delete this user" })
+        }
     }
     catch(error){
         console.log(error)
